Updated formatting to separate logic from slash commands
